Refactored the backend structure to centralize route management, and fixed an issue where the super admin could not be correctly identified during registration.
