+++ .env.0example
@@ -17,9 +17,9 @@ MAIL_USERNAME=yourgmail@gmail.com
 MAIL_PASSWORD=mailpassword
 
 #Prime administrator
-ADMIN_USERNAME=your_username
-ADMIN_PASSWORD=your_password
-ADMIN_EMAIL=your_email
+SUPERADMIN_USERNAME=your_username
+SUPERADMIN_PASSWORD=your_password
+SUPERADMIN_EMAIL=your_email
 
 #Test
 TEST_USERNAME=Test
