+++ Backend/routes/auth/avatar.py
