+++ docker-compose.1dev.yml
@@ -1,4 +1,3 @@
-version: '3.9'
 services:
   frontend:
     build:
