+++ docker-compose.3prod.yml
@@ -1,4 +1,3 @@
-version: "3.9"
 services:
   frontend:
     image: jerryzheng0907/angular-frontend:latest
